+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Change Log <a id="Change_Log"></a>
 *  Version 1.2a
     * Reformatted WTS/WTB command output to be more compact.
     * New !botstats (!bs) command to show misc. bot statistics
+    * New !feedback command to leave comments/suggestions for the developer.
 *  Version 1.1.3a
     * Added !alch command for finding out the high alch value of items using their RS2007 item names
     * Improved Hiscore parsing reliability